Fix project structure and update license
I moved project files from the nested 'monero-rust-miner' directory to the repository root.
This addresses the issue of incorrect project placement and an unnecessary level of directory nesting.

Key changes:
- All contents of the 'monero-rust-miner' directory (source code, Cargo files, README, etc.) are now at the root level.
- I removed the now-empty 'monero-rust-miner' directory.
- I replaced the placeholder root README.md with the detailed project README.md.
- I ensured the correct 'modules' directory (from the original project) is at the root.
- I updated the LICENSE file content from Apache 2.0 to MIT License to match the README.md description. Placeholders [year] and [fullname] were used for the copyright line in the MIT license.
