+++ modules/mining_engine/randomx_wrapper.h
