+++ modules/mining_engine/src/tests/engine_tests.rs
