--- modules/rpc_connector/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "monero-rpc-connector"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
