--- modules/rpc_connector/src/lib.rs
@@ -1,51 +1,0 @@
-// Placeholder for rpc_connector lib
-pub struct RpcConnector;
-impl RpcConnector {
-    pub async fn new(_config: RpcConfig) -> Result<Self, anyhow::Error> {
-        Ok(RpcConnector)
-    }
-    pub async fn get_block_template(&self, _wallet_address: &str) -> Result<MiningJob, anyhow::Error> {
-        Ok(MiningJob::default())
-    }
-    pub async fn submit_block(&self, _block_blob: &str) -> Result<String, anyhow::Error> {
-        Ok("OK".to_string())
-    }
-    pub async fn start_job_fetch_loop(
-        &self,
-        _job_tx: tokio::sync::mpsc::Sender<MiningJob>,
-        _cancellation_tx: tokio::sync::watch::Sender<u64>,
-        _shutdown_rx: tokio::sync::broadcast::Receiver<()>,
-    ) -> Result<(), anyhow::Error> {
-        Ok(())
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct RpcConfig {
-    pub url: String,
-    pub username: Option<String>,
-    pub password: Option<String>,
-    pub wallet_address: String,
-    pub check_interval_secs: u64,
-}
-
-#[derive(Clone, Debug, Default)]
-pub struct MiningJob {
-    pub block_template_blob: String,
-    pub difficulty: u64,
-    pub height: u64,
-    pub reserved_offset: u32,
-    pub prev_hash: String,
-    pub seed_hash: String,
-    pub job_id: String,
-}
-
-// Define MoneroRpc trait
-pub trait MoneroRpc {
-    // Define associated types and methods as needed
-}
-
-// Implement MoneroRpc for RpcConnector
-impl MoneroRpc for RpcConnector {
-    // Implement trait methods
-}
