--- monero-rust-miner/modules/mining_engine/Cargo.toml
@@ -1,49 +1,0 @@
-# monero-rust-miner/modules/mining_engine/Cargo.toml
-[package]
-name = "monero-mining-engine"
-version = "0.1.0"
-edition = "2021"
-authors = ["Your Name <your.email@example.com>"] # ЗАМЕНИТЬ
-description = "High-performance RandomX mining engine for Monero."
-license = "MIT"
-
-[dependencies]
-# Асинхронный рантайм и утилиты
-tokio = { version = "1.45.1", features = ["macros", "rt-multi-thread", "sync", "time"] }
-
-# Обработка ошибок
-anyhow = "1.0.98"
-
-# Логирование
-tracing = "0.1.41"
-
-# Параллельные вычисления на CPU
-rayon = "1.10.0"
-
-# Работа с большими числами (для wide_difficulty)
-num-bigint = "0.4.6"
-num-traits = "0.2.19"
-
-# Работа с HEX-строками
-hex = "0.4.3"
-
-# Сериализация/десериализация (для MinerConfig)
-serde = { version = "1.0.219", features = ["derive"] }
-
-# Для FFI и работы с C-типами
-libc = "0.2.153"
-# is_x86_feature_detected! макрос находится в std::arch, поэтому отдельный крейт не всегда нужен,
-# но libc может быть полезен для других C-типов, если bindgen их не покроет.
-
-# Зависимость от нашего RPC-коннектора для типа MiningJob
-monero-rpc-connector = { path = "../../modules/rpc_connector" }
-
-[build-dependencies]
-# Скрипт build.rs использует `cc` для сборки RandomX и `bindgen` для генерации FFI-биндингов.
-cc = "1.2.25"
-bindgen = "0.71.1"
-
-[features]
-default = []
-# Фича для build.rs: отключает -march=native для переносимости.
-portable_build = []
