--- monero-rust-miner/modules/mining_engine/src/.gitkeep
@@ -1,1 +1,0 @@
-# This file is intentionally left blank to ensure the directory is tracked by Git.
